+++ src/Service/ExportCustomerData/ExportCustomerDataService.php
@@ -18,7 +18,7 @@
  * @license   https://opensource.org/licenses/AFL-3.0 Academic Free License version 3.0
  */
 
-namespace PrestaShop\Module\Psgdpr\Service\Export;
+namespace PrestaShop\Module\Psgdpr\Service\ExportCustomerData;
 
 use Cart;
 use CartRule;
